requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
SQLAlchemy==2.0.36
PyJWT==2.9.0
bcrypt==4.2.0
reportlab==4.2.5
python-docx==1.1.2
python-dotenv==1.0.1
waitress==3.0.0

# Optional for DAST only:
# frida==16.5.9
# frida-tools==13.6.0
